Evangelism, Discipleship, Missions
EDM is a Christian
organization in Sierra Leone,
West Africa that heard the
call of evangelism,
discipleship, and missions in
that part of the world.

● How it all started:
○ Years ago God put a burden on my heart for
issues I noticed in the church.
○ The church was focused inwardly
(programs, budgets, activities), while
neglecting others outside.
○ Ephesians 2:3-4

● There was no formal training structure for
new converts to grow in maturity in the faith.
● The purpose of God for our salvation is to
grow to become like Christ Jesus
● Matthew 28:18-20
● When I look at the church, priority is not given
to evangelism.
● Romans 10:13-15

● As I was studying and praying I began to form
the organization and came up with the
structure of EDM.
● Evangelism, Discipleship, and Missions

Mission Statement: In our acts of obedience and
worship to God, we will go out and share the good
news to the unbelievers, train them to maturity so

they will train others.

Motto: Love God and love others.

● Vision for EDM:
○ Build a campus for the body of Christ
■ Retreat center, school, lodging, recreation,
chapel, etc.
○ Short term goals:
■ Get the van and equipment through
customs
■ Finalize land acquisition
■ Establish international board members
■ Begin construction of school
■ Evangelism training
■ Show the Jesus movie for evangelism

● Long term goals:
○ Build facilities on the EDM campus.
■ Bible school
■ Retreat center/meeting hall
○ Extend the mission to other cities
○ Establish departments:
■ Education
■ Training
■ Outreach/missions
■ Development/projects
■ Church ministry
■ Building and maintenance

● Updates:
○ We have acquired land to build a school and a
retreat center for the Body of Christ.
○ The van and musical instruments shipped in
August, and will arrive the first week of
October.
○ There is an outstanding balance of $3,500.00
to pay the customs in Freetown.

● How to help:
○ I want to take this opportunity to thank all
those who have generously contributed toward
this venture.
○ I want to appeal for financial support to
accomplish all this.
○ Please consider donating one time, monthly,
or quarterly.