i the about page, this statement was made by the visioner EDWIN KARGBO
Years ago, God put a burden on my heart for issues I noticed in the church. The church was often focused inwardly—on programs, budgets, and activities—while sometimes neglecting those outside its walls. This concern resonated with Ephesians 2:3-4, which reminds us of God's rich mercy and great love.

I observed a lack of formal training structures for new converts to grow in maturity in the faith. Yet, the purpose of God for our salvation is to grow to become like Christ Jesus. The Great Commission in Matthew 28:18-20 commands us to make disciples and teach them.

Looking at the church, it seemed that evangelism was not always given the priority it deserved, despite the call in Romans 10:13-15.

As I studied, prayed, and reflected on these challenges, the vision for EDM began to form. The organization was structured around three core pillars: Evangelism, Discipleship, and Missions.